wircrc: fix incorrect handling of wmk race
old bug; unlike original version which wrote cmsg's to all windows at
once, we spawn a window for that channel or user if it doesn't exist,
but we almost always end up sending more text for the window wmk
hasn't registered yet.  the stupid wins2 array was used for it, the
intent apparently being to just resend the message until wmk has
called wname and is done, but wins2 is never updated, so this spins at
100% cpu.  instead, use a pointer variable and background a proc which
at least sleeps while trying to resend the data, instead of slowing
down the main rc thread, and doesn't loop.